+++ pycrest/eve.py
@@ -1,10 +1,6 @@
-import os
 import base64
 import requests
 import time
-import zlib
-import hashlib
-import pickle
 from pycrest import version
 from pycrest.compat import bytes_, text_
 from pycrest.errors import APIException, UnsupportedHTTPMethodException
@@ -15,132 +11,15 @@ except ImportError:  # pragma: no cover
     from urlparse import urlparse, urlunparse, parse_qsl
 
 try:
-    import pickle
-except ImportError:  # pragma: no cover
-    import cPickle as pickle
-
-try:
     from urllib.parse import quote
 except ImportError:  # pragma: no cover
     from urllib import quote
 import logging
 import re
+from pycrest.cache import DictCache, APICache, DummyCache
 
 logger = logging.getLogger("pycrest.eve")
 cache_re = re.compile(r'max-age=([0-9]+)')
-
-
-class APICache(object):
-
-    def put(self, key, value):
-        raise NotImplementedError
-
-    def get(self, key):
-        raise NotImplementedError
-
-    def invalidate(self, key):
-        raise NotImplementedError
-
-    def _hash(self, data):
-        h = hashlib.new('md5')
-        h.update(pickle.dumps(data))
-        #prefix allows possibility of multiple applications 
-        #sharing same keyspace
-        return 'pyc_' + h.hexdigest()
-
-
-class FileCache(APICache):
-
-    def __init__(self, path):
-        self._cache = {}
-        self.path = path
-        if not os.path.isdir(self.path):
-            os.mkdir(self.path, 0o700)
-
-    def _getpath(self, key):
-        return os.path.join(self.path, self._hash(key) + '.cache')
-
-    def put(self, key, value):
-        with open(self._getpath(key), 'wb') as f:
-            f.write(
-                zlib.compress(
-                    pickle.dumps(value,
-                                 pickle.HIGHEST_PROTOCOL)))
-        self._cache[key] = value
-
-    def get(self, key):
-        if key in self._cache:
-            return self._cache[key]
-
-        try:
-            with open(self._getpath(key), 'rb') as f:
-                return pickle.loads(zlib.decompress(f.read()))
-        except IOError as ex:
-            logger.debug('IOError: {0}'.format(ex))
-            if ex.errno == 2:  # file does not exist (yet)
-                return None
-            else:   # pragma: no cover
-                raise
-
-    def invalidate(self, key):
-        self._cache.pop(key, None)
-
-        try:
-            os.unlink(self._getpath(key))
-        except OSError as ex:
-            if ex.errno == 2:  # does not exist
-                pass
-            else:   # pragma: no cover
-                raise
-
-
-class DictCache(APICache):
-
-    def __init__(self):
-        self._dict = {}
-
-    def get(self, key):
-        return self._dict.get(key, None)
-
-    def put(self, key, value):
-        self._dict[key] = value
-
-    def invalidate(self, key):
-        self._dict.pop(key, None)
-
-        
-class DummyCache(APICache):
-    """ Provide a fake cache class to allow a "no cache"
-        use without breaking everything within APIConnection """
-    def __init__(self):
-        self._dict = {}
-
-    def get(self, key):
-        return None
-
-    def put(self, key, value):
-        pass
-
-    def invalidate(self, key):
-        pass
-        
-
-class MemcachedCache(APICache):
-
-    def __init__(self, serverList):  # serverList could be ['127.0.0.1:11211']
-        # import memcache here so that the dependency on the python-memcached
-        # only occurs if you use it
-        import memcache
-        self._mc = memcache.Client(serverList, debug=0)
-
-    def get(self, key):
-        return self._mc.get(self._hash(key))
-
-    def put(self, key, value):
-        return self._mc.set(self._hash(key), value)
-
-    def invalidate(self, key):
-        return self._mc.delete(self._hash(key))
 
 
 class APIConnection(object):
@@ -149,8 +28,15 @@ class APIConnection(object):
             self,
             additional_headers=None,
             user_agent=None,
-            cache_dir=None,
-            cache=DictCache):
+            **kwargs):
+        '''Initialises a PyCrest object
+
+        Keyword arguments:
+        additional_headers - a list of http headers that will be sent to the server
+        user_agent - a custom user agent
+        cache - an instance of an APICache object that will cache HTTP Requests.
+                Default is DictCache, pass cache=None to disable caching.
+        '''
         # Set up a Requests Session
         session = requests.Session()
         if additional_headers is None:
@@ -164,16 +50,16 @@ class APIConnection(object):
         })
         session.headers.update(additional_headers)
         self._session = session
-        if cache_dir:
-            self.cache_dir = cache_dir
-            self.cache = FileCache(self.cache_dir)
-        elif cache:
-            if isinstance(cache, APICache):
-                self.cache = cache  # Inherit from parents
-            elif isinstance(cache, type):
-                self.cache = cache()  # Instantiate a new cache
+        if 'cache' not in kwargs:
+            self.cache = DictCache()
         else:
-            self.cache = DummyCache()
+            cache = kwargs.pop('cache')
+            if isinstance(cache, APICache):
+                self.cache = cache
+            elif cache is None:
+                self.cache = DummyCache()
+            else:
+                raise ValueError('Provided cache must implement APICache')
 
     def _parse_parameters(self, resource, params):
         '''Creates a dictionary from query_string and `params`
